% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inf-series.R
\name{inf_series}
\alias{inf_series}
\title{A function to sum an infinite series}
\usage{
inf_series(a, r)
}
\arguments{
\item{a}{the first item of the infinite series}

\item{r}{the ratio of the subsequent items in the series}
}
\description{
This function calculates the sum of an infinite series.
}
